07.03 - Validando no controller

+++ app/Http/Controllers/ProjectController.php
@@ -27,7 +27,7 @@ class ProjectController extends Controller
      */
     public function index()
     {
-        return $this->repository->all();
+        return $this->repository->findWhere(['owner_id'=>\Authorizer::getResourceOwnerId()]);
     }
 
     /**
@@ -49,6 +49,9 @@ class ProjectController extends Controller
      */
     public function show($id)
     {
+        if(!$this->checkProjectOwner($id)){
+            return ['error' => 'Acesso Negado'];
+        }
         return $this->repository->find($id);
     }
 
@@ -61,6 +64,9 @@ class ProjectController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(!$this->checkProjectOwner($id)){
+            return ['error' => 'Acesso Negado'];
+        }
         return $this->service->update($request->all(), $id);
     }
 
@@ -72,6 +78,16 @@ class ProjectController extends Controller
      */
     public function destroy($id)
     {
+        if(!$this->checkProjectOwner($id)){
+            return ['error' => 'Acesso Negado'];
+        }
         return $this->repository->delete($id);
+    }
+    
+    private function checkProjectOwner($projectId) 
+    {
+        $userId = \Authorizer::getResourceOwnerId();
+        
+        return $this->repository->isOwner($projectId,$userId);
     }
 }
